JUnit test cases.
Each test case will use their own Spring Boot instance. Thus the Spring Boot
will start multiple times in the same process to server each test case.

To verify the Spring Boot could restart multiple times, we implemented
two test cases in the. Added the annotation @DirtiesContext to base test case
to solve the issue that when restart the Spring Boot in the same process, it
report that the port 8080 already in use.